# Improve launch performance of kernels