Before:

```
[ Info: Ka Launch
BenchmarkTools.Trial: 
  memory estimate:  1.73 KiB
  allocs estimate:  32
  --------------
  minimum time:     9.702 μs (0.00% GC)
  median time:      10.228 μs (0.00% GC)
  mean time:        11.063 μs (0.00% GC)
  maximum time:     68.124 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

```

After:
```
[ Info: Ka Launch
BenchmarkTools.Trial: 
  memory estimate:  816 bytes
  allocs estimate:  9
  --------------
  minimum time:     4.653 μs (0.00% GC)
  median time:      5.074 μs (0.00% GC)
  mean time:        5.444 μs (0.57% GC)
  maximum time:     317.282 μs (97.24% GC)
  --------------
  samples:          10000
  evals/sample:     7

```