@vchuravy 2020-03-27
Becnhmark script from #75 

```
using BenchmarkTools
using KernelAbstractions

@kernel function update_ka!(dQ, Q, rka, rkb, dt)
  i = @index(Global, Linear)
  @inbounds begin
    Q[i] += rkb * dt * dQ[i]
    dQ[i] *= rka
  end
end

dQ = zeros(1)
Q = rand(1)
rka = rand()
rkb = rand()
dt = rand()
kernel = update_ka!(CPU(), 256, 1)

kal = @benchmark(event = $kernel($dQ, $Q, $rka, $rkb, $dt), setup=(event=nothing), teardown=(wait(event)))
@info "Ka Launch"
display(kal)
println()
```